crypto-pgpainless: extract utility methods to `KeyUtils`
